Move survey as the last item on body, add cookie configuration that makes it possible to siteimprove ignore the survey, refactor accordion code so that reusable code is available for the survey code as well
