+++ src/js/accordion/accordion-state.js
@@ -1,37 +1,23 @@
+import ClientHelpers from '../client-helpers';
 import LocalStorageManager from '../localStorageManager';
 
-export default class State {
+export default class AccordionState {
 
   constructor() {
     this.storageManager = new LocalStorageManager('helfi-settings');
-
     this.site = window.drupalSettings.helfi_instance_name || '';
     this.page = window.drupalSettings.path.currentPath;
-    this.siteAccordionStates = JSON.parse(this.storageManager.getValue(this.getStorageKey())) || {};
+    this.siteAccordionStates = JSON.parse(this.storageManager.getValue(this.getAccordionStorageKey())) || {};
     this.pageAccordionStates = this.siteAccordionStates[this.page] || {};
 
     // Initialize the cookie check. This check is for Siteimprove so that it can
     // have all the accordions open always so that it can read the contents.
-    this.AccordionsOpen = State.isCookieSet('helfi_accordions_open');
+    this.AccordionsOpen = ClientHelpers.isCookieSet('helfi_accordions_open');
   }
 
-  static isCookieSet = cookieName => {
-    const cookies = document.cookie;
-    const cookieArray = cookies.split('; ');
+  getAccordionStorageKey = () => `${this.site}-accordion`;
 
-    // Loop through the cookies to see if desired one is set.
-    for (let i = 0; i < cookieArray.length; i++) {
-      const cookiePair = cookieArray[i].split('=');
-      if (cookiePair[0] === cookieName) {
-        return true;
-      }
-    }
-    return false;
-  };
-
-  getStorageKey = () => `${this.site}-accordion`;
-
-  saveItemState = (accordionItemId, isOpen) => {
+  saveAccordionItemState = (accordionItemId, isOpen) => {
     if (!this.site) return false;
     if (!this.siteAccordionStates[this.page]) this.siteAccordionStates[this.page] = {};
     // Save only the open accordion items to the local storage.
@@ -40,15 +26,14 @@ export default class State {
     } else {
       this.siteAccordionStates[this.page][accordionItemId] = isOpen;
     }
-    this.storageManager.setValue(this.getStorageKey(), JSON.stringify(this.siteAccordionStates));
+    this.storageManager.setValue(this.getAccordionStorageKey(), JSON.stringify(this.siteAccordionStates));
   };
 
-  loadItemState = accordionItemId => {
+  loadAccordionItemState = accordionItemId => {
     // Use the cached cookie check result
     if (this.AccordionsOpen) return true;
     if (!this.site) return false;
     return !!this.pageAccordionStates[accordionItemId];
   };
 
-  static getCurrentLanguage = () => window.drupalSettings.path.currentLanguage;
 }
